Grant unfiltered_html to administrator role
In Wordpress multisite, only super admins
have this capability. This is because adding
unfiltered HTML could be abused to hijack
a superadmin's account via a XSS attack or similar.

In WMCZ's case, this is not a concern. Admins are
highly trustworthy (generally either staff members
or trustees), and their list is auditted regularly.
Adding this capability should not result in significant
added risks.

This is needed to update certain forms at the web.
Since we do not need this capability at the blog,
let's do this for web only (principle of least privilege).

Change-Id: Id87b3361d8be2a48df8e23cb885650c7a457a423